requirements.txt
aiogram==2.14
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asyncpg==0.29.0
attrs==23.2.0
Babel==2.15.0
certifi==2024.6.2
click==8.1.7
colorama==0.4.6
diff-match-patch==20230430
Django==5.0.6
django-extensions==3.2.3
django-import-export==4.0.8
dnspython==2.6.1
email_validator==2.1.1
environs==8.0.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-jwt-auth==0.5.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
multidict==6.0.5
orjson==3.10.4
packaging==24.1
pillow==10.3.0
psycopg2==2.9.9
pydantic==1.10.12
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.30
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
starlette==0.37.2
tablib==3.5.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
yarl==1.9.4